base-files/flash: add possibility to flash with /tmp and /generic- drives
/media/hdd and /media/usb should be working too if removed.
Drives like usb sticks or hdd's must have the right
drive name ''usb'' or ''hdd'' and this is not always available.
An empty and unnamed stick will be detected as Generic- and will be
used for flashing. If no drive is available it will fall back
to /tmp. I've tested and it has worked fine at many tries.
For experienced users is it always a good idea to use a stick or
hard drive, but with right names!