base-files: add patch number to mark different original recipe
This change ensures the creation of a new base file package and allows
for updates without requiring a complete machine flash, even when only
modified or new additional base files are added.

Previously, the package creation was not consistently performed
(for unknown reasons) when changes were made to this recipe.
As a result, updates for this package were not fetched from
the package feeds on the machine. Instead, it was necessary
to first execute "bitbake -c cleansstate base-files" and rebuild and
flash a complete image. By increasing this patch number, this is now
enforced. Additionally, the version information of this package clearly
indicates that it is an extension of the original recipe, as it has been
augmented with a ".bbappend" file.